# src/microservices/proxy/api/openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: support@cinemaabyss.com
    name: CinemaAbyss Team
  description: |
    API спецификация для системы CinemaAbyss, включающая монолит и микросервисы.

    Система CinemaAbyss представляет собой платформу для управления фильмами, пользователями, платежами и подписками.
    Архитектура системы включает монолитное приложение и выделенные микросервисы.
  title: CinemaAbyss API
  version: 1.0.0
servers:
- description: API Gateway (Proxy Service)
  url: http://localhost:8000
- description: Монолитное приложение
  url: http://localhost:8080
- description: Микросервис фильмов
  url: http://localhost:8081
- description: Микросервис событий
  url: http://localhost:8082
...
paths:
...
  /api/events/movie:
    post:
      description: "Регистрирует новое событие, связанное с фильмом"
      operationId: createMovieEvent
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MovieEvent"
        description: Данные события фильма
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventResponse"
          description: Событие успешно создано
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Некорректный запрос
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Внутренняя ошибка сервера
      summary: Создание события фильма
      tags:
      - events
  /api/events/user:
    post:
      description: "Регистрирует новое событие, связанное с пользователем"
      operationId: createUserEvent
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserEvent"
        description: Данные события пользователя
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventResponse"
          description: Событие успешно создано
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Некорректный запрос
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Внутренняя ошибка сервера
      summary: Создание события пользователя
      tags:
      - events
  /api/events/payment:
    post:
      description: "Регистрирует новое событие, связанное с платежом"
      operationId: createPaymentEvent
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentEvent"
        description: Данные события платежа
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventResponse"
          description: Событие успешно создано
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Некорректный запрос
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Внутренняя ошибка сервера
      summary: Создание события платежа
      tags:
      - events
components:
  schemas:
...
    Event:
      example:
        id: movie-1-viewed
        type: movie
        timestamp: 2023-01-15T14:30:00Z
        payload: "{}"
      properties:
        id:
          description: Уникальный идентификатор события
          example: movie-1-viewed
          type: string
        type:
          description: Тип события
          example: movie
          type: string
        timestamp:
          description: Время события
          example: 2023-01-15T14:30:00Z
          format: date-time
          type: string
        payload:
          description: Полезная нагрузка события (зависит от типа события)
          type: object
      required:
      - id
      - payload
      - timestamp
      - type
      type: object
    MovieEvent:
      example:
        movie_id: 1
        title: Inception
        action: viewed
        user_id: 1
        rating: 8.5
        genres:
        - Sci-Fi
        - Action
        description: A mind-bending thriller
      properties:
        movie_id:
          description: Идентификатор фильма
          example: 1
          type: integer
        title:
          description: Название фильма
          example: Inception
          type: string
        action:
          description: Действие с фильмом
          example: viewed
          type: string
        user_id:
          description: Идентификатор пользователя (опционально)
          example: 1
          type: integer
        rating:
          description: Рейтинг (опционально)
          example: 8.5
          format: float
          type: number
        genres:
          description: Жанры фильма (опционально)
          example:
          - Sci-Fi
          - Action
          items:
            type: string
          type: array
        description:
          description: Описание фильма (опционально)
          example: A mind-bending thriller
          type: string
      required:
      - action
      - movie_id
      - title
      type: object
    UserEvent:
      example:
        user_id: 1
        username: john_doe
        email: john.doe@example.com
        action: registered
        timestamp: 2023-01-15T14:30:00Z
      properties:
        user_id:
          description: Идентификатор пользователя
          example: 1
          type: integer
        username:
          description: Имя пользователя (опционально)
          example: john_doe
          type: string
        email:
          description: Email пользователя (опционально)
          example: john.doe@example.com
          type: string
        action:
          description: Действие пользователя
          example: registered
          type: string
        timestamp:
          description: Время события
          example: 2023-01-15T14:30:00Z
          format: date-time
          type: string
      required:
      - action
      - timestamp
      - user_id
      type: object
    PaymentEvent:
      example:
        payment_id: 1
        user_id: 1
        amount: 9.99
        status: completed
        timestamp: 2023-01-15T14:30:00Z
        method_type: credit_card
      properties:
        payment_id:
          description: Идентификатор платежа
          example: 1
          type: integer
        user_id:
          description: Идентификатор пользователя
          example: 1
          type: integer
        amount:
          description: Сумма платежа
          example: 9.99
          format: float
          type: number
        status:
          description: Статус платежа
          example: completed
          type: string
        timestamp:
          description: Время платежа
          example: 2023-01-15T14:30:00Z
          format: date-time
          type: string
        method_type:
          description: Тип метода оплаты (опционально)
          example: credit_card
          type: string
      required:
      - amount
      - payment_id
      - status
      - timestamp
      - user_id
      type: object
    EventResponse:
      example:
        status: success
        partition: 0
        offset: 42
        event:
          id: movie-1-viewed
          type: movie
          timestamp: 2023-01-15T14:30:00Z
          payload: "{}"
      properties:
        status:
          description: Статус операции
          example: success
          type: string
        partition:
          description: Партиция Kafka
          example: 0
          type: integer
        offset:
          description: Смещение в партиции Kafka
          example: 42
          type: integer
        event:
          $ref: "#/components/schemas/Event"
      required:
      - event
      - offset
      - partition
      - status
      type: object
    Error:
      example:
        error: Internal Server Error
      properties:
        error:
          description: Сообщение об ошибке
          example: Internal Server Error
          type: string
      required:
      - error
      type: object
...