# src/microservices/proxy/go/model_movie_event.go
// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

/*
 * CinemaAbyss API
 *
 * API спецификация для системы CinemaAbyss, включающая монолит и микросервисы.  Система CinemaAbyss представляет собой платформу для управления фильмами, пользователями, платежами и подписками. Архитектура системы включает монолитное приложение и выделенные микросервисы. 
 *
 * API version: 1.0.0
 * Contact: support@cinemaabyss.com
 */

package openapi




type MovieEvent struct {

	// Идентификатор фильма
	MovieId int32 `json:"movie_id"`

	// Название фильма
	Title string `json:"title"`

	// Действие с фильмом
	Action string `json:"action"`

	// Идентификатор пользователя (опционально)
	UserId int32 `json:"user_id,omitempty"`

	// Рейтинг (опционально)
	Rating float32 `json:"rating,omitempty"`

	// Жанры фильма (опционально)
	Genres []string `json:"genres,omitempty"`

	// Описание фильма (опционально)
	Description string `json:"description,omitempty"`
}

// AssertMovieEventRequired checks if the required fields are not zero-ed
func AssertMovieEventRequired(obj MovieEvent) error {
	elements := map[string]interface{}{
		"movie_id": obj.MovieId,
		"title": obj.Title,
		"action": obj.Action,
	}
	for name, el := range elements {
		if isZero := IsZeroValue(el); isZero {
			return &RequiredError{Field: name}
		}
	}

	return nil
}

// AssertMovieEventConstraints checks if the values respects the defined constraints
func AssertMovieEventConstraints(obj MovieEvent) error {
	return nil
}