# src/microservices/proxy/go/model_payment_event.go
// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

/*
 * CinemaAbyss API
 *
 * API спецификация для системы CinemaAbyss, включающая монолит и микросервисы.  Система CinemaAbyss представляет собой платформу для управления фильмами, пользователями, платежами и подписками. Архитектура системы включает монолитное приложение и выделенные микросервисы. 
 *
 * API version: 1.0.0
 * Contact: support@cinemaabyss.com
 */

package openapi


import (
	"time"
)



type PaymentEvent struct {

	// Идентификатор платежа
	PaymentId int32 `json:"payment_id"`

	// Идентификатор пользователя
	UserId int32 `json:"user_id"`

	// Сумма платежа
	Amount float32 `json:"amount"`

	// Статус платежа
	Status string `json:"status"`

	// Время платежа
	Timestamp time.Time `json:"timestamp"`

	// Тип метода оплаты (опционально)
	MethodType string `json:"method_type,omitempty"`
}

// AssertPaymentEventRequired checks if the required fields are not zero-ed
func AssertPaymentEventRequired(obj PaymentEvent) error {
	elements := map[string]interface{}{
		"payment_id": obj.PaymentId,
		"user_id": obj.UserId,
		"amount": obj.Amount,
		"status": obj.Status,
		"timestamp": obj.Timestamp,
	}
	for name, el := range elements {
		if isZero := IsZeroValue(el); isZero {
			return &RequiredError{Field: name}
		}
	}

	return nil
}

// AssertPaymentEventConstraints checks if the values respects the defined constraints
func AssertPaymentEventConstraints(obj PaymentEvent) error {
	return nil
}