# src/microservices/proxy/go/model_user_event.go
// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

/*
 * CinemaAbyss API
 *
 * API спецификация для системы CinemaAbyss, включающая монолит и микросервисы.  Система CinemaAbyss представляет собой платформу для управления фильмами, пользователями, платежами и подписками. Архитектура системы включает монолитное приложение и выделенные микросервисы. 
 *
 * API version: 1.0.0
 * Contact: support@cinemaabyss.com
 */

package openapi


import (
	"time"
)



type UserEvent struct {

	// Идентификатор пользователя
	UserId int32 `json:"user_id"`

	// Имя пользователя (опционально)
	Username string `json:"username,omitempty"`

	// Email пользователя (опционально)
	Email string `json:"email,omitempty"`

	// Действие пользователя
	Action string `json:"action"`

	// Время события
	Timestamp time.Time `json:"timestamp"`
}

// AssertUserEventRequired checks if the required fields are not zero-ed
func AssertUserEventRequired(obj UserEvent) error {
	elements := map[string]interface{}{
		"user_id": obj.UserId,
		"action": obj.Action,
		"timestamp": obj.Timestamp,
	}
	for name, el := range elements {
		if isZero := IsZeroValue(el); isZero {
			return &RequiredError{Field: name}
		}
	}

	return nil
}

// AssertUserEventConstraints checks if the values respects the defined constraints
func AssertUserEventConstraints(obj UserEvent) error {
	return nil
}